Simplify activiy in share SDK
